tools: add a protocol debugger
This adds a Python-based protocol-debugger that can be used as MITM
between a server and clients. The easiest invocation of it is
to run without argument on the same host as the server and then connect
to it via a client on port 24801.

This will simply print out all known messages as they fly between client
and server, making it easier to figure out what's going on on the wire.

Note that the message parsing was intentionally hand-implemented to
potentially show any discrepancies between implementations that rely
on the format strings. The format strings from ProtocolTypes.cpp are
included for human verification of the parsing vs what is expected.

Typical output is:

10:15:56 - deskflow        - I - Listening on port 24801, connecting to localhost:24800
10:15:56 - deskflow        - I - Client connected from 192.168.1.235:54952
10:15:56 - deskflow        - I - Connected to remote localhost:24800
10:15:56 - server → client - I - Barrier major=1, minor=8
10:15:56 - server ← client - I - Barrier major=1, minor=8
10:15:56 - server → client - I - QINF
10:15:56 - server → client - I - LSYN data=en
10:15:56 - server ← client - I - DINF x=0, y=0, w=1920, h=1080, mx=0, my=960, size=540
10:15:56 - server → client - I - CIAK
10:15:56 - server → client - I - CROP
10:15:56 - server → client - I - DSOP options=30
[...]
10:16:48 - server ← client - I - CNOP
10:16:48 - server → client - I - DMMV x=897, y=297
10:16:48 - server ← client - I - CNOP
10:16:48 - server → client - I - DMMV x=895, y=295
10:16:48 - server ← client - I - CNOP
10:16:48 - server → client - I - DKDL keyid=97, mask=0, button=38, lang= (obfuscated)
10:16:48 - server ← client - I - CNOP
10:16:48 - server → client - I - DKDL keyid=97, mask=0, button=38, lang= (obfuscated)

By default keycodes and buttons are obfuscated to always 97/38 (the 'a'
key). This is to avoid users accidentally leaking private data when
attaching logs, use --show-keycodes if debugging key values is actually
neccessary.